refactor: modularize worker social auth flow
